Update patch set 11
Patch Set 11:

(1 comment)

> Patch Set 10:
> 
> There are some issues detected by static analysis. See "scan_build" folder.
In the static analyzer steps, steps 11 and 12 are mutually exclusive, because, for any child request, the request->parent_request field will be populated,and goes into the else branch in step 12. Only if the request is a single request or parent request, then only the if condition in step 12 goes into true branch. Hence this is not a valid flow of execution in this case.

Patch-set: 11
Reviewer: Rakesh.Babu@wdc.com <1013920@d5d70762-12d0-45a1-890d-524b12d3f735>