Update patch set 54
Patch Set 54: Code-Review-1

(4 comments)

This is awesome!  This gives me a lot of confidence that we will be able to deprecate the old config file format sometime soon.

I haven't reviewed the config converter Python script in detail.  The tests look good to me though.

Just a few comments - mostly on the commit message and the CHANGELOG.  This is a really important patch in my opinion so want to make sure we have a good commit message and CHANGELOG entry for it.

Patch-set: 54
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1