Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> It's still not clear to me what you're trying to say in this patch. Here are some scenarios:
> 
> 1) The user has their IOMMU enabled. They run scripts/setup.sh, which automatically selects vfio-pci. That works fine.
> 
> 2) The user has their IOMMU disabled. They run scripts/setup.sh, which automatically selects uio_pci_generic. That works fine.
> 
> 3) The user has their IOMMU enabled, but not configured. They manually bind devices to uio_pci_generic. That *should* work fine, but I haven't tested it.
> 
> 4) The user has their IOMMU enabled and configured to protect their PCI device. They manually bind their device to uio_pci_generic. That will not work and is a user error.
> 
> Which one of these four scenarios are you trying to clarify? In practice, users will only ever see scenarios 1 and 2 and as far as I'm aware those are both working correctly.

Hi Ben,

Thanks a lot! The scenarios are much clear. :)

Case 1,2 are the normal cases, case 3, 4 are the abnormal cases. The change is used to guide the user to check their environment when met troubles. For case 1, 2 and 3, SPDK works; for case 4, SPDK doesn't work, and no error log output, 

The patch is used to clarify the issue in the case 3 and 4.

Thanks!

Patch-set: 2
Reviewer: tone.zhang@arm.com <1014914@d5d70762-12d0-45a1-890d-524b12d3f735>