Update patch set 1
Patch Set 1:

I don't think it makes sense to add a way to send an IO without a timeout at all. There has to be a timeout for any request, even if the timeout value is very large (may happen in some odd cases). An infinite wait for a request is not a likely intention of the user.

Patch-set: 1
CC: Baruch Even <1014422@d5d70762-12d0-45a1-890d-524b12d3f735>