Update patch set 1
Patch Set 1:

> Hi Vitaliy,
Hi, Jim!

 > 
 > When you ran this on your Ubuntu server, did you run it as root, or
 > as yourself?
 > 
 > On my system, if I run the trap command under /bin/sh as myself, I
 > get the same error message.  But if I run it under /bin/sh as root,
 > it is fine.

I tried both root and unprivileged user, but got the same error.

 > 
 > It really would be better to not force people to use bash here.   
 > It would be better to make a $USER = root check similar to the one in
 > scripts/vagrant/update.sh.
 > 
 > -Jim

Since root didn't work out for me, I suspect that cause of this problem is that on most Debian-like distros /bin/sh is pointing to dash (it was the case for mine Ubuntu as well) and dash does not support ERR SIGNAL_SPEC which was used by the script.
I agree that forcing bash on people is bad thing, but there seems to be no easy solution for this problem if we need to use trap with ERR SIGNAL_SPEC.

Also, here is useful info about ERR:
https://en.wikibooks.org/wiki/Bourne_Shell_Scripting/Debugging_and_signal_handling#Err..._ERR.3F
 > ERR isn't standard at all. It was added by the Korn Shell to make life easier, but not adopted by the POSIX standard and it certainly isn't part of the original Bourne Shell. So if you use it, remember that your script may not be portable anymore.

Patch-set: 1
Reviewer: Vitaliy Mysak <1014935@d5d70762-12d0-45a1-890d-524b12d3f735>