Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> can you briefly explain what problem this is solving or what additional benefit it adds?

This change is related to how we register memory with the NIC in RDMA. Up until now, memory has been registered as a single region inside the NIC and we haven't had to worry much about an I/O spanning multiple MRs inside of the NIC. However, that mode will be changing as we start to use DPDKs dynamic memory allocation. In this case, if we cross an MR boundary, we won't be able to submit the I/O without splitting it.
This is the first step in trying to enable splitting I/O accross multiple MRs. The next step will include allowing us to send multiple SGL elements, one for each MR if we end up splitting.

Patch-set: 4
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>