Update patch set 2
Patch Set 2: Code-Review-1

Hi,

Thanks for your explanations behind this patch.

What if we modified the code to:

1) check if device is attached to uio_pci_generic
2) check if IOMMU is enabled

If #1 and #2 are both true, print an error message.

I think this is possible.  My concern with the changes to memory.md is that new users will not read this document, and even if they did read this document, it would not explain exactly what the problem is.

-Jim

Patch-set: 2
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1