Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> I think we should keep 'size' but for a different reason.
> 
> We should use 'size' to make sure the translated address is valid for the full region requested by the user.
> 
> So I'd just drop this patch for now.  We can add the checks for the full range later.

My next patch adds in a check for that at the rdma level. I was initially going to implement length checking inside of spdk_mem_map_translate, but the maps treat this condition differently.
Say a user requests a 4MB region. In each map, you would have to retrieve the values for two translation keys. But having a valid translation for the full 4MB looks different for each map.
1. vtophys. A valid (contiguous) 4 mb translation would return 2 distinct values addr1 and addr2=addr1 + 2MB for the two keys.
2. RDMA mr map. A valid translation would return the same value (i.e. a pointer to the same mr struct) for both keys.
I can't think of a consistent way of checking if a translation is valid beyond a single mem_map entry inside of this function. That's why I got rid of size here and moved the check up a level.
Unless you mean that a user passes in a memory address in the middle of a 2MB page and we want to pass back some value from 1 - 2MB that informs the user how much runway they have from that address to the beginning of the next page (or map entry)?

Patch-set: 1
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>