iscsi: Remove LUN field check that cause SCSI compliance test failure
A major SCSI compliance test expects that the LUN field is reserved
(not to be checked) in the SCSI Data-Out PDU.

In that test, the LUN field in the first SCSI Command PDU is 0 and
the LUN field in the SCSI Data-Out PDU sent in response to the R2T
is 0xFFFFFFFFFFFFFFFF.

Hence spdk_iscsi_op_data() returns Check Condition because LUN is
not found.

This patch removes LUN field check in spdk_iscsi_op_data().
The LUN field is only checked and doesn't change any behavior.
Hence it can be removed safely.

Change-Id: I110a7e396d1e517b2a39ca5e586ab2bb2d45e5f3
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>