env/dpdk: do not merge adjacent registrations
This prevents us from unregistering two or more separately
registered regions with a single notification.

This works for the current workflow, but there are still
various potential unhandled cases when memory is registered
more than once. One suggestion to handle those is to forbid
registering memory twice.

Unregistering a part of a previously registered region is
still possible and is handled in a separate patch.

Change-Id: I7d61867fa61fd7a4a8a644ff45cab17125d63e1b
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>