scripts: remove unused dependency
test/common/config/vm_setup.sh installs package 'perl-open' which is unused.
It is the problem especially if someone is setting up tests on Debian-based machine because it is hard to find equivalent package.
There seems to be single script which intended to use perl-open, namely test/app/match/match, but it works without it.

Change-Id: I949452ec786f1574e02813226313becf5327b03d
Signed-off-by: Vitaliy Mysak <vitaliy.mysak@intel.com>