Update patch set 2
Patch Set 2:

> I tried to review the patch series. But Lance's feedback is
 > difficult for me to score yet. Sorry for back-and-force.

Hi Shuhei.   I intentionally provided no score (0) because I felt the solution is less than optimal since it results in artificially creating dependencies for a number of SPDK shared libs which aren't required.   However, the patch does specifically address a problem with at least one of the SPDK libs --   libspdk_nvmf.so -- and in doing so resolves github issue 434.   If a SPDK developer builds the SPDK from sources and does not configure with RDMA, then the libibverbs and librdmacm will effectively be removed from the build so the size of the lib dependencies is significantly reduced leaving just a couple dependencies that likely will be required as a minimal set needed for SPDK apps.   And, so, I'm on the fence.   I can be persuaded to go along with this patch, but I wanted to highlight this point for other reviews to weigh in.

Patch-set: 2