env/dpdk: ensure memory is unregistered in chunks it was registered with
In order to control how the memory is registered and unregistered,
we will now control how it is allocated. We need to ensure all
dynamically added hugepages will be unregistered in the same chunks
they were registered. The only problematic case here is when a single
hugepage was registered as a part of bigger region (>2MB allocation),
but was later reused to hold smaller allocations. The scenario is as
follows (assuming 2MB pages):
 * allocate 3MB of memory -> 2 pages are reserved and registered
                             as one region
 * allocate 500KB of memory -> one page from the previous allocation
                               is reused
 * free 3MB -> one of the pages still keeps the 500KB allocation,
               so only the other page is released and unregistered
To prevent the case above, we extend each allocation bigger than 2MB
(reduced by cache line size to make place for dpdk metadata) to
occupy entire pages.

Buffers smaller than 2MB and crossing the hugepage boundary do not
cause any trouble as they partially reuse a page that's already
registered, hence cause only another single page to be dynamically
allocated and registered.

spdk_ring and spdk_mempool still need to be refactored, but as of
now they are never freed at runtime so they don't cause any trouble.

Change-Id: I08b8afd751a095be6e618ddf47a93c2b0e9fe5c4
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>