Update patch set 1
Patch Set 1: Code-Review-1

I don't think we want to do this.  Removing O_DIRECT will add page caching into the kernel path which I don't think we want to do.  It requires us to implement flushing and other checks to make sure that data written is *actually* written to disk.

I think the underlying problem here is that at least some code paths in the VM submit I/O buffers that aren't aligned on LBA-sized boundaries.  This will never happen with iSCSI or NVMe-oF since we allocate the buffers in those cases.  But for vhost we are dependent on the buffers provided by the VM.

I think we need to try to detect mis-aligned buffers in the AIO driver.  If they are unaligned, allocate an aligned buffer, copy the data, then write.  Or allocate an aligned buffer, read, then copy the data.

Hopefully this is a rare case.  The submitter already said this generally works on Linux.  It's possible that on Windows, we only have corner cases (i.e. userspace apps with unaligned malloc'd buffers) that hit this case.

Patch-set: 1
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1