Update patch set 24
Patch Set 24: Code-Review-1

We need to save the io_unit in the superblock for backward compatibility.  I think we talked about this previously.  With this patch, a logical volume on a previously existing blobstore on a 512-byte LBA device will now report a 512-byte LBA instead of 4KB.  I don't think it's OK to implicitly change the block size like this.

So I think we need to save io_unit_size in the superblock.  Use reserved space.  If it's set to zero (like it will be for older blobstores), then assume 4KB.

Patch-set: 24
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1