Update patch set 2
Patch Set 2:

I'm reverting the last suggested change to vbdev_crypto_dump_info_json() as it would require me to put a parm of some kind in the write message.  Looking at all other bdev modules dump routines, none of them are using the write_named_object as was suggested in the review. It's not really clear to me when/why/what the difference is in some of the json write routines, will get some more education on that later but for now this is how all the others are written. If this is a real error please let me know

Patch-set: 2
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>