Update patch set 3
Patch Set 3:

(4 comments)

I made some comments, but some of them don't need to get fixed in this particular patch. In general I'd prefer patches to fix just one thing at a time instead of being a collection of miscellaneous cleanups. One rule of thumb I use for myself is that if I start making a list of items in the commit message, I should probably break the patch up. So in that spirit, if you do decide to address some of my larger comments, do them in separate patches that follow this one.

Patch-set: 3
CC: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>