Update patch set 31
Patch Set 31:

(6 comments)

The code looks really good. I have a few nits but nothing major. 

It should be noted that all the comments are relative to 18.07. When you switch to 18.10, you'll have to move to the new message queues for both io and control events, but that change should be relatively simple, as long as you maintain the synchronous approach to reading/writing (i.e., as long as you don't need things like epoll).

Patch-set: 31
CC: Florin Coras <1015094@d5d70762-12d0-45a1-890d-524b12d3f735>