Update patch set 32
Patch Set 32:

(3 comments)

> Patch Set 31:
> 
> (6 comments)
> 
> The code looks really good. I have a few nits but nothing major. 
> 
> It should be noted that all the comments are relative to 18.07. When you switch to 18.10, you'll have to move to the new message queues for both io and control events, but that change should be relatively simple, as long as you maintain the synchronous approach to reading/writing (i.e., as long as you don't need things like epoll).

Thank you for your review.

For now we're using posix-like model in non-blocking mode to handle connections and we're not able to improve the code in some areas without changing an API of our library (made it more asynchronous, improve socket/session pools and so on). Also data and command flow should probably be changed, but it will require much more work on our libsock library as well as application which uses it. In this step we don't want to made so far changes, but we will definitely use your advises in the next step.

Patch-set: 32
Reviewer: Tomasz Kulasek <1013008@d5d70762-12d0-45a1-890d-524b12d3f735>