Update patch set 9
Patch Set 9: Code-Review-1

I think this may be too specific to RDMA. For RDMA, we definitely want to make the unregister notifications match the register notifications, because they can only be unregistered in those units. For regular memory (vtophys) though, we don't want to do that.

I have a few general thoughts that I'll outline, but I'm not entirely sure what the right answer is. First, I'm not sure we need to reference count the memory registrations anymore. Is this something that can occur in a non-error scenario? Maybe it is - I'm not sure. If we convince ourselves that we don't, we can remove that entire extra reference counting map.

Second, I think we do need something like what you've implemented here, but we need to do it only within the RDMA layer. Basically, the RDMA layer needs to create two memory maps - one that tracks where the MRs are and one that tracks the translation of va to rkey. Then we can allow unregistrations of rkey regions and remove them from that map, but only unregister the MR if it is entirely released.

Patch-set: 9
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1