Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This is probably a stupid question but shouldn't the sourcing of the config file come after the new assignments you added starting at line 74?

That's a good question. The code starting on line 74 is, as Daniel used to call it "syntactic black magic". There's no way of knowing what it does just by looking at it. What each of those lines means translates to is "if x is not set in the environment, then set it to value. Then unconditionally export x". So we have to source the configuration file to give right of first refusal to any user defined configurations, then if the user doesn't define any one of the above variables, we fill in the blanks.
I think we could technically source the file afterwards, but this way is valid too and follows the pattern we use in autotest_common.sh

Patch-set: 1
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>