Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Hmm, I didn't know you could git clone a specific branch, cool.  So before we were just grabbing the latest? This is better for a few reasons then!

Well we were cloning from the tip of our fork of nvme-cli so it wasn't exactly bleeding edge code. I think that Daniel checked out a release and applied a few patches to it. I just did the same thing with the 1.6 release and put it on a new branch. I did it that way to keep a consistent pattern with how we do forked repos. (See our rocksdb branches for example).

Patch-set: 1
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>