Update patch set 5
Patch Set 5: Code-Review-1

After building with this patch + 426131, an "objdump -t" against the combined shared lib, libspdk.so, suggests that some symbols are missing.   For example, "objdump -t | grep nvmf" yielded nothing.   A build without 426129+426131 followed by an inspection of the symbols in libspdk.so does show nvmf symbols.   A compare of the actual compile lines, though, on just the libspdk.so.1.0 for the "good" case (i.e. symbols present) vs the case with missing symbols looks identical.   However, the size of the library that is built is very different.   In the good case, it's about 5-6x the size of the one that is missing symbols.

Patch-set: 5
Reviewer: Lance Hartmann <1011995@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1