Update patch set 5
Patch Set 5: -Code-Review

Further inspection of the libspdk.so.1.0 with ldd reveals the difference between the versions pre and post patches.   The smaller one with "missing symbols" is due to the fact that the lib is not constructed from static libs, but instead is relying on dependencies to other spdk shared libs.   So, they may resolve after all.   Going to try a couple of experiments to sanity check.

Patch-set: 5
Reviewer: Lance Hartmann <1011995@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0