Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> After building with this patch + 426131, an "objdump -t" against the combined shared lib, libspdk.so, suggests that some symbols are missing.   For example, "objdump -t | grep nvmf" yielded nothing.   A build without 426129+426131 followed by an inspection of the symbols in libspdk.so does show nvmf symbols.   A compare of the actual compile lines, though, on just the libspdk.so.1.0 for the "good" case (i.e. symbols present) vs the case with missing symbols looks identical.   However, the size of the library that is built is very different.   In the good case, it's about 5-6x the size of the one that is missing symbols.

Just posting a reply here to summarize what we figured out on IRC. The reason that the shared library is much smaller now is because instead of including the entirety of each of the included libraries, it is linking to the already compiled individual shared objects. This results in not duplicating the symbols between the individual and combined shared objects.

Patch-set: 5
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>